# backend/src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Taska API
  description: |
    Taska is a real-time tasking platform connecting customers with local
    taskers for delivery and errand services.

    ## Authentication

    Most endpoints require a JWT access token obtained via the
    `POST /api/v1/auth/verify-otp` flow.

    Include the token in the `Authorization` header:
    ```
    Authorization: Bearer <accessToken>
    ```

    The token contains the user's `userId`, `phoneNumber`, and `role` (CUSTOMER | TASKER | ADMIN).
  version: 1.0.0
  contact:
    name: Taska Support
    url: https://taska.app
servers:
  - url: /api/v1
    description: API v1
paths:
...
  /wallet:
    get:
      tags: [Wallet]
      summary: Get wallet
      description: |
        Returns the wallet for the authenticated tasker.
      operationId: getWallet
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Wallet details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Tasker profile not found
          $ref: '#/components/responses/NotFound'
  /wallet/balance:
    get:
      tags: [Wallet]
      summary: Get balance summary
      description: |
        Returns a summary of the tasker's wallet balances including
        current balance, pending earnings, available balance,
        and withdrawable amount.
      operationId: getBalanceSummary
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Balance summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Tasker profile not found
          $ref: '#/components/responses/NotFound'
  /wallet/transactions:
    get:
      tags: [Wallet]
      summary: Get transaction history
      description: |
        Returns the append-only transaction history for the tasker's wallet.
      operationId: getTransactions
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Transaction history
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/WalletTransaction'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  /wallet/withdraw:
    post:
      tags: [Wallet]
      summary: Request withdrawal
      description: |
        Requests a withdrawal from the tasker's available balance.
      operationId: requestWithdrawal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                  description: Amount to withdraw in ETB
            example:
              amount: 300.00
      responses:
        '200':
          description: Withdrawal processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: number
                    description: Remaining wallet balance
                  withdrawn:
                    type: number
                    description: Amount withdrawn
        '400':
          description: Insufficient balance
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Wallet not found
          $ref: '#/components/responses/NotFound'
  '/wallet/admin/{taskerId}':
    get:
      tags: [Wallet]
      summary: Get wallet by tasker ID (Admin)
      description: |
        Admin-only endpoint. Returns the wallet for any tasker.
      operationId: getWalletByTaskerId
      security:
        - bearerAuth: []
      parameters:
        - name: taskerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Tasker profile ID
      responses:
        '200':
          description: Wallet details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Admin access required
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─────────────────────────────────────────────────────
  #  Webhooks
  # ─────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `POST /api/v1/auth/verify-otp`.
        Include as `Authorization: Bearer <token>`.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Human-readable error message
      required: [success, error]

    # ── Health ────────────────────────────────────────────
...
    Wallet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        taskerId:
          type: string
          format: uuid
        balance:
          type: number
          format: double
          description: Total wallet balance
        pendingBalance:
          type: number
          format: double
          description: Unsettled earnings
        availableBalance:
          type: number
          format: double
          description: Balance available for withdrawal
        totalEarned:
          type: number
          format: double
        totalWithdrawn:
          type: number
          format: double
        totalRefunded:
          type: number
          format: double
        currency:
          type: string
          default: ETB
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, taskerId, balance, pendingBalance, availableBalance, totalEarned, totalWithdrawn, totalRefunded, currency, createdAt, updatedAt]
    WalletResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Wallet'
      required: [success, data]
    BalanceSummary:
      type: object
      properties:
        balance:
          type: number
        pendingBalance:
          type: number
        availableBalance:
          type: number
        totalEarned:
          type: number
        totalWithdrawn:
          type: number
        totalRefunded:
          type: number
        withdrawable:
          type: number
          description: Amount currently available for withdrawal
        currency:
          type: string
    WalletTransaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [TASK_EARNING, WITHDRAWAL, REFUND, ADJUSTMENT, DEPOSIT, PAYMENT, FEE, PAYOUT, CREDIT, DEBIT, HOLD, RELEASE]
        amount:
          type: number
          format: double
        description:
          type: string
          nullable: true
        referenceId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time

    # ── Reusable websocket schemas (informational) ────────
...
  responses:
    Unauthorized:
      description: Authentication required — missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Authentication required
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Invalid folder. Allowed folders: profile-images, task-images, verification-documents, vehicle-images"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Resource not found
...